app/requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.7
django-filter==24.3
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-database-url==2.2.0
drf-yasg==1.21.7
importlib_resources==6.4.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
PyJWT==2.9.0
PyYAML==6.0.1
packaging==24.1
paystackapi==2.1.3
psycopg2-binary
python-decouple==3.8
pytz==2024.1
referencing==0.35.1
rpds-py==0.19.1
sqlparse==0.5.1
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.7.0